# backend/dist/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserResponseType": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"editor"
						]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"name",
					"email",
					"role",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"PersonResponseType": {
				"description": "Representa a estrutura de dados de uma pessoa (colaborador, escrito etc.) \r\nretornada pela API.",
				"properties": {
					"_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"kind": {
						"type": "string"
					},
					"description": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"contact": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"_id",
					"name",
					"kind",
					"description",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreatePersonInput": {
				"properties": {
					"contact": {
						"type": "string"
					},
					"description": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"kind": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Inferência de tipos a partir dos schemas Zod.\r\nGarante que as mutações (Post/Put) sigam estritamente o contrato definido no Zod."
			},
			"UpdatePersonInput": {
				"properties": {
					"contact": {
						"type": "string"
					},
					"description": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"kind": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"type": "object"
			},
...
			"LoginResponseType": {
				"description": "Representa a estrutura de resposta enviada ao cliente após uma autenticação bem-sucedida.",
				"properties": {
					"user": {
						"$ref": "#/components/schemas/UserResponseType",
						"description": "* Dados do perfil do usuário autenticado (geralmente omitindo a senha por segurança)."
					},
					"token": {
						"type": "string",
						"description": "* Token de acesso (geralmente JWT) que deve ser enviado nas próximas requisições \r\ndentro do header 'Authorization: Bearer <token>'."
					}
				},
				"required": [
					"user",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginInput": {
				"properties": {
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"RequestResetInput": {
				"properties": {
					"email": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"ResetPasswordInput": {
				"properties": {
					"newPassword": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					}
				},
				"type": "object"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Token JWT no formato 'Bearer {token}'"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "Backend para o TCC Usina Guara",
		"contact": {}
	},
	"paths": {
...
		"/people": {
			"post": {
				"operationId": "CreatePerson",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PersonResponseType"
								}
							}
						}
					}
				},
				"description": "Registra uma nova pessoa no sistema.\r\nRequer autenticação JWT.",
				"tags": [
					"People"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Dados da pessoa (validados via Zod)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePersonInput",
								"description": "Dados da pessoa (validados via Zod)"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "FindPeople",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/PersonResponseType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Recupera uma lista de pessoas.",
				"tags": [
					"People"
				],
				"security": [],
				"parameters": [
					{
						"description": "Filtro opcional para categorizar o tipo de pessoa (ex: 'colaborador', 'cliente')",
						"in": "query",
						"name": "kind",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/people/{id}": {
			"get": {
				"operationId": "GetPersonById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PersonResponseType"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Busca os detalhes de uma pessoa específica pelo ID único.",
				"tags": [
					"People"
				],
				"security": [],
				"parameters": [
					{
						"description": "Identificador UUID ou ID do banco de dados",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdatePerson",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PersonResponseType"
								}
							}
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Atualiza os dados de uma pessoa existente.\r\nRequer autenticação JWT.",
				"tags": [
					"People"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID da pessoa a ser editada",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdatePersonInput"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeletePerson",
				"responses": {
					"204": {
						"description": "No Content"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"description": "Remove permanentemente uma pessoa do registro.\r\nRequer autenticação JWT.",
				"tags": [
					"People"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "ID da pessoa a ser excluída",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LoginResponseType"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized - Credenciais inválidas"
					}
				},
				"description": "Autentica um usuário e retorna um token de acesso.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Objeto contendo as credenciais (email e password)",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginInput",
								"description": "Objeto contendo as credenciais (email e password)"
							}
						}
					}
				}
			}
		},
		"/auth/request-password-reset": {
			"post": {
				"operationId": "RequestPasswordReset",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "E-mail não encontrado"
					}
				},
				"description": "Inicia o fluxo de recuperação de senha enviando um e-mail com instruções/token.",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Objeto contendo o e-mail do usuário",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestResetInput",
								"description": "Objeto contendo o e-mail do usuário"
							}
						}
					}
				}
			}
		},
		"/auth/reset-password": {
			"post": {
				"operationId": "ResetPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"description": "Define uma nova senha para o usuário utilizando o codigo padrao",
				"tags": [
					"Authentication"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "Objeto contendo o token e a nova senha",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ResetPasswordInput",
								"description": "Objeto contendo o token e a nova senha"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}